+++ Microting.EformBackendConfigurationBase/Migrations/20220103152353_AddingComplianceStatusToProperty.Designer.cs
@@ -692,7 +692,7 @@ namespace Microting.EformBackendConfigurationBase.Migrations
                     b.Property<int>("AreaRulePlanningId")
                         .HasColumnType("int");
 
-                    b.Property<int>("AreaRulesId")
+                    b.Property<int>("AreaRuleId")
                         .HasColumnType("int");
 
                     b.Property<DateTime>("CreatedAt")